+++ lib/attributes_and_token_lists.rb
@@ -1,12 +1,12 @@
 require "attributes_and_token_lists/version"
 require "attributes_and_token_lists/engine"
-require "attributes_and_token_lists/attributes_builder"
+require "attributes_and_token_lists/tag_builder"
 
 module AttributesAndTokenLists
   mattr_accessor(:config) { ActiveSupport::OrderedOptions.new }
 
   def self.builder(name, &block)
-    instance = config.builders[name] = Class.new(AttributesBuilder)
+    instance = config.builders[name] = Class.new(TagBuilder)
 
     if block.present?
       block.arity.zero? ? instance.instance_exec(&block) : instance.yield_self(&block)